+++ backend/types/types.go
@@ -3,7 +3,7 @@ package types
 import "github.com/go-playground/validator/v10"
 
 type User struct {
-	ID        uint64 `json:"id" validate:"required"`
+	ID        uint   `json:"id" validate:"required"`
 	Username  string `json:"username" validate:"required"`
 	GithubURL string `json:"git_url" validate:"required"`
 }
@@ -11,7 +11,7 @@ type User struct {
 type Repo struct {
 	Name    string `json:"name" validate:"required"`
 	Private bool   `json:"private" validate:"required"`
-	UserID  int    `json:"user_id" validate:"required"`
+	UserID  uint   `json:"user_id" validate:"required"`
 }
 
 type File struct {
